add map caculate

+++ detect/run.py
@@ -1,10 +1,11 @@
-from .common import models_path, num_model, folder_path, read_jpg_files, config_path, output_file
+from .common import models_path, num_model, folder_path, read_jpg_files, config_path, output_file, test_dir
 from .threads import YoloThread, CodetrThread
 from .confirm import confirm
+from .test import get_map
 from detect.softmax.outfeature import output_dataset
 from detect.softmax.train import train
 from detect.softmax.makeset import mk_set, cln_set
-from detect.softmax.tempfile import train_img_dir, train_data, val_data
+from detect.softmax.tempfile import train_img_dir, train_data, val_data, train_set, val_set
 from ultralytics import YOLO
 from mmdet.apis import init_detector
 import os
@@ -17,7 +18,7 @@ def run(mode='detect'):
             mode 用于切换模式，
                 其中detect是运行检测，最终得到的是检测结果
                 train_mmodel是用于训练中间件模型
-
+                map 用于得出mAP分析结果
     """
     with open(output_file, 'w') as file:
         file.write('')
@@ -28,6 +29,9 @@ def run(mode='detect'):
         file_list = read_jpg_files(folder_path)
     elif mode == 'train_mmodel':
         file_list = read_jpg_files(train_img_dir)
+    elif mode == "mAP":
+        file_list = read_jpg_files(test_dir)
+        all_results = []
 
     for i in range(num_model):
         if config_path[i] == '':
@@ -52,6 +56,11 @@ def run(mode='detect'):
             confirm(os.path.basename(file), single_results)
         elif mode == 'train_mmodel':
             output_dataset(os.path.basename(file), single_results)
+        elif mode == "mAP":
+            all_results.append((os.path.basename(file), confirm(os.path.basename(file), single_results, "mAP")))
+
+    if mode == "mAP":
+        get_map(all_results)
 
 
 def train_mmodels():
@@ -61,12 +70,16 @@ def train_mmodels():
         except OSError as exc:  # 防止并发创建目录时出错
             if exc.errno != os.errno.EEXIST:
                 raise
-    with open(train_data, 'w') as train_out:
-        with open(val_data, 'w') as val_out:
-            train_out.write('')
-            val_out.write('')
 
-    run(mode='train_mmodel')
-    mk_set()
-    cln_set()
+    if not (os.path.exists(train_data) and os.path.exists(val_data)):
+        with open(train_data, 'w') as train_out:
+            with open(val_data, 'w') as val_out:
+                train_out.write('')
+                val_out.write('')
+        run(mode='train_mmodel')
+
+    if not (os.path.exists(train_set) and os.path.exists(val_set)):
+        mk_set()
+        cln_set()
+
     train()
